refactor: make SearchInput reusable outside /susun
